+++ src/main/java/noppes/npcs/attribute/player/PlayerAttribute.java
@@ -6,9 +6,9 @@ import noppes.npcs.attribute.IAttributeInstance;
 
 public class PlayerAttribute implements IAttributeInstance {
     private final AttributeDefinition attribute;
-    private double baseValue;
+    private float baseValue;
 
-    public PlayerAttribute(AttributeDefinition attribute, double baseValue) {
+    public PlayerAttribute(AttributeDefinition attribute, float baseValue) {
         this.attribute = attribute;
         this.baseValue = baseValue;
     }
@@ -19,12 +19,12 @@ public class PlayerAttribute implements IAttributeInstance {
     }
 
     @Override
-    public double getValue() {
+    public float getValue() {
         return baseValue;
     }
 
     @Override
-    public void setValue(double value) {
+    public void setValue(float value) {
         this.baseValue = value;
     }
 }
