+++ src/main/java/noppes/npcs/attribute/player/PlayerAttributeTracker.java
@@ -14,10 +14,10 @@ public class PlayerAttributeTracker {
     private PlayerAttributeMap playerAttributes = new PlayerAttributeMap();
 
     public final int extraHealth = 0;
-    public final Map<Integer, Double> magicDamage = new HashMap<>();
-    public final Map<Integer, Double> magicBoost = new HashMap<>();
-    public final Map<Integer, Double> magicDefense = new HashMap<>();
-    public final Map<Integer, Double> magicResistance = new HashMap<>();
+    public final Map<Integer, Float> magicDamage = new HashMap<>();
+    public final Map<Integer, Float> magicBoost = new HashMap<>();
+    public final Map<Integer, Float> magicDefense = new HashMap<>();
+    public final Map<Integer, Float> magicResistance = new HashMap<>();
 
     private PlayerEquipmentTracker equipmentTracker = new PlayerEquipmentTracker();
 
@@ -46,14 +46,14 @@ public class PlayerAttributeTracker {
             ItemStack[] equipment = new ItemStack[] { currentEquip.heldItem, currentEquip.boots, currentEquip.leggings, currentEquip.chestplate, currentEquip.helmet };
             for (ItemStack piece : equipment) {
                 if (piece != null) {
-                    for (Entry<String, Double> entry : ItemAttributeHelper.readAttributes(piece).entrySet()) {
+                    for (Entry<String, Float> entry : ItemAttributeHelper.readAttributes(piece).entrySet()) {
                         String key = entry.getKey();
-                        double value = entry.getValue();
+                        float value = entry.getValue();
                         AttributeDefinition def = AttributeController.getAttribute(key);
                         if (def != null) {
                             IAttributeInstance inst = playerAttributes.getAttributeInstance(def);
                             if (inst == null) {
-                                inst = playerAttributes.registerAttribute(def, 0.0);
+                                inst = playerAttributes.registerAttribute(def, 0.0f);
                             }
                             inst.setValue(inst.getValue() + value);
                         }
@@ -61,31 +61,31 @@ public class PlayerAttributeTracker {
                     // Process Magic
 
                     // OFFENSIVE MAGIC
-                    Map<Integer, Double> magicFlat = ItemAttributeHelper.readMagicAttributeMap(piece, ModAttributes.MAGIC_DAMAGE_KEY);
-                    for (Entry<Integer, Double> entry : magicFlat.entrySet()) {
+                    Map<Integer, Float> magicFlat = ItemAttributeHelper.readMagicAttributeMap(piece, ModAttributes.MAGIC_DAMAGE_KEY);
+                    for (Entry<Integer, Float> entry : magicFlat.entrySet()) {
                         int magicId = entry.getKey();
-                        double value = entry.getValue();
-                        magicDamage.put(magicId, magicDamage.getOrDefault(magicId, 0.0) + value);
+                        float value = entry.getValue();
+                        magicDamage.put(magicId, magicDamage.getOrDefault(magicId, 0.0f) + value);
                     }
-                    Map<Integer, Double> magicPercent = ItemAttributeHelper.readMagicAttributeMap(piece, ModAttributes.MAGIC_BOOST_KEY);
-                    for (Entry<Integer, Double> entry : magicPercent.entrySet()) {
+                    Map<Integer, Float> magicPercent = ItemAttributeHelper.readMagicAttributeMap(piece, ModAttributes.MAGIC_BOOST_KEY);
+                    for (Entry<Integer, Float> entry : magicPercent.entrySet()) {
                         int magicId = entry.getKey();
-                        double value = entry.getValue();
-                        magicBoost.put(magicId, magicBoost.getOrDefault(magicId, 0.0) + value);
+                        float value = entry.getValue();
+                        magicBoost.put(magicId, magicBoost.getOrDefault(magicId, 0.0f) + value);
                     }
 
                     // DEFENSIVE MAGIC
-                    Map<Integer, Double> magicDefFlat = ItemAttributeHelper.readMagicAttributeMap(piece, ModAttributes.MAGIC_DEFENSE_KEY);
-                    for (Entry<Integer, Double> entry : magicDefFlat.entrySet()) {
+                    Map<Integer, Float> magicDefFlat = ItemAttributeHelper.readMagicAttributeMap(piece, ModAttributes.MAGIC_DEFENSE_KEY);
+                    for (Entry<Integer, Float> entry : magicDefFlat.entrySet()) {
                         int magicId = entry.getKey();
-                        double value = entry.getValue();
-                        magicDefense.put(magicId, magicDefense.getOrDefault(magicId, 0.0) + value);
+                        float value = entry.getValue();
+                        magicDefense.put(magicId, magicDefense.getOrDefault(magicId, 0.0f) + value);
                     }
-                    Map<Integer, Double> magicResist = ItemAttributeHelper.readMagicAttributeMap(piece, ModAttributes.MAGIC_RESISTANCE_KEY);
-                    for (Entry<Integer, Double> entry : magicResist.entrySet()) {
+                    Map<Integer, Float> magicResist = ItemAttributeHelper.readMagicAttributeMap(piece, ModAttributes.MAGIC_RESISTANCE_KEY);
+                    for (Entry<Integer, Float> entry : magicResist.entrySet()) {
                         int magicId = entry.getKey();
-                        double value = entry.getValue();
-                        magicResistance.put(magicId, magicResistance.getOrDefault(magicId, 0.0) + value);
+                        float value = entry.getValue();
+                        magicResistance.put(magicId, magicResistance.getOrDefault(magicId, 0.0f) + value);
                     }
                 }
             }
@@ -104,8 +104,8 @@ public class PlayerAttributeTracker {
         }
     }
 
-    public double getAttributeValue(AttributeDefinition def) {
+    public float getAttributeValue(AttributeDefinition def) {
         IAttributeInstance inst = playerAttributes.getAttributeInstance(def);
-        return inst != null ? inst.getValue() : 0.0;
+        return inst != null ? inst.getValue() : 0.0f;
     }
 }
